cleanup unused code, fixed ts errors
