+++ .github/workflows/migrate-dev-to-main.yml
@@ -12,7 +12,7 @@ jobs:
 
       - name: Load Secrets in Env
         run: |
-          npx tsx ./scripts/load-secrets
+          npx tsx ./scripts/load-secrets-ci
         env:
           NODE_ENV: dev
           INFISICAL_CLIENT_ID: ${{ secrets.INFISICAL_CLIENT_ID }}
@@ -23,7 +23,7 @@ jobs:
         run: |
           npm install -g @xata.io/cli@latest
 
-      - name: Start Xata Migration from dev to main
+      - name: Migrate DB schema from dev to main
         run: |
           npx tsx ./scripts/xata-migration dev main
         env:
