+++ .github/workflows/migrate-main-to-dev.yml
@@ -12,7 +12,7 @@ jobs:
 
       - name: Load Secrets in Env
         run: |
-          npx tsx ./scripts/load-secrets
+          npx tsx ./scripts/load-secrets-ci
         env:
           NODE_ENV: production
           INFISICAL_CLIENT_ID: ${{ secrets.INFISICAL_CLIENT_ID }}
@@ -23,7 +23,7 @@ jobs:
         run: |
           npm install -g @xata.io/cli@latest
 
-      - name: Start Xata Migration from main to dev
+      - name: Migrate DB schema from main to dev
         run: |
           npx tsx ./scripts/xata-migration main dev
         env:
