+++ .github/workflows/reload-secrets.yml
@@ -19,7 +19,7 @@ jobs:
     steps:
       - name: Load Secrets in Env
         run: |
-          npx tsx ./scripts/load-secrets
+          npx tsx ./scripts/load-secrets-ci
         env:
           NODE_ENV: ${{ github.event.inputs.environment }}
           INFISICAL_CLIENT_ID: ${{ secrets.INFISICAL_CLIENT_ID }}
