+++ .xata/version/compatibility.json
@@ -1,1 +1,10 @@
-{"@xata.io/cli":{"latest":"0.16.11","compatibility":[{"range":">=0.0.0"}]},"@xata.io/client":{"latest":"0.30.1","compatibility":[{"range":">=0.0.0"}]}}
+{
+  "@xata.io/cli": {
+    "latest": "0.16.11",
+    "compatibility": [{ "range": ">=0.0.0" }]
+  },
+  "@xata.io/client": {
+    "latest": "0.30.1",
+    "compatibility": [{ "range": ">=0.0.0" }]
+  }
+}
